new: usr: allow multiple handlers for same event with sorting using order.
Got rid of the "scope" positional argument when registering an event.
It was used to allow multiple handlers for one event, overcoming
Dectate's conflict resolution. You can now pass in nothing for the
first registration and it defaults to 20. After that, you have to
provide an "order" value to disambiguate, which also allows ordering
of the handlers. 40-80 (on each event type) are reserved for the
system.